Make net::IOBuffer 64-bit safe
BUG=166496

Review URL: https://chromiumcodereview.appspot.com/12091071

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@179730 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/base/io_buffer.h
@@ -116,7 +116,7 @@ class NET_EXPORT StringIOBuffer : public IOBuffer {
  public:
   explicit StringIOBuffer(const std::string& s);
 
-  int size() const { return string_data_.size(); }
+  int size() const { return static_cast<int>(string_data_.size()); }
 
  private:
   virtual ~StringIOBuffer();
